Further refactoring issue #44 Renamed NullComparator to NullConditionalCheck moved ConditionalCheck interface to comparison package moved ConditionalCheckResult to result package
