+++ src/main/java/org/testmonkeys/jentitytest/comparison/Comparator.java
@@ -1,7 +1,6 @@
 package org.testmonkeys.jentitytest.comparison;
 
 import org.testmonkeys.jentitytest.comparison.result.ComparisonResult;
-import org.testmonkeys.jentitytest.comparison.util.ConditionalCheck;
 import org.testmonkeys.jentitytest.framework.JEntityTestException;
 
 import java.beans.PropertyDescriptor;
