+++ src/main/java/org/testmonkeys/jentitytest/comparison/MultiResultComparator.java
@@ -1,8 +1,7 @@
 package org.testmonkeys.jentitytest.comparison;
 
 import org.testmonkeys.jentitytest.comparison.result.ComparisonResult;
-import org.testmonkeys.jentitytest.comparison.util.ConditionalCheck;
-import org.testmonkeys.jentitytest.comparison.util.ConditionalCheckResult;
+import org.testmonkeys.jentitytest.comparison.result.ConditionalCheckResult;
 import org.testmonkeys.jentitytest.framework.JEntityTestException;
 
 import java.beans.PropertyDescriptor;
