+++ src/main/java/org/testmonkeys/jentitytest/comparison/entity/ChildEntityComparator.java
@@ -2,8 +2,8 @@ package org.testmonkeys.jentitytest.comparison.entity;
 
 import org.testmonkeys.jentitytest.comparison.ComparisonContext;
 import org.testmonkeys.jentitytest.comparison.MultiResultComparator;
+import org.testmonkeys.jentitytest.comparison.conditionalChecks.NullConditionalCheck;
 import org.testmonkeys.jentitytest.comparison.result.ComparisonResult;
-import org.testmonkeys.jentitytest.comparison.util.NullComparator;
 import org.testmonkeys.jentitytest.framework.JEntityTestException;
 
 import java.util.List;
@@ -12,7 +12,7 @@ import java.util.List;
 public class ChildEntityComparator extends MultiResultComparator {
 
     public ChildEntityComparator() {
-        registerPreConditionalCheck(new NullComparator());
+        registerPreConditionalCheck(new NullConditionalCheck());
     }
 
     @Override
