+++ src/test/java/org/testmonkeys/jentitytest/test/unit/entityInspectionTests/models/BadComparator.java
@@ -2,8 +2,8 @@ package org.testmonkeys.jentitytest.test.unit.entityInspectionTests.models;
 
 import org.testmonkeys.jentitytest.comparison.Comparator;
 import org.testmonkeys.jentitytest.comparison.ComparisonContext;
+import org.testmonkeys.jentitytest.comparison.ConditionalCheck;
 import org.testmonkeys.jentitytest.comparison.result.ComparisonResult;
-import org.testmonkeys.jentitytest.comparison.util.ConditionalCheck;
 import org.testmonkeys.jentitytest.framework.JEntityTestException;
 
 import java.beans.PropertyDescriptor;
